Airflow/requirements.txt
requests
PyPDF2
apache-airflow
python-dotenv
bcrypt
boto3
lxml
sqlalchemy